requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
Flask-Compress==1.14
Flask-Cors==4.0.0
redis==5.0.1
aiohttp==3.9.1
google-generativeai==0.3.1
bcrypt==4.1.2
PyJWT==2.8.0
cryptography==41.0.7
python-dotenv==1.0.0
gunicorn==21.2.0
